numpy type set to int for parsing

+++ xas/trajectory.py
@@ -183,8 +183,8 @@ class trajectory():
             postedge_hi = edge_energy + offsets[3]
             edge = edge_energy
 
-            x1_num = np.round(half * 100)
-            x2_num = np.round((1 - half) * 100)
+            x1_num = int(np.round(half * 100))
+            x2_num = int(np.round((1 - half) * 100))
             x1 = np.linspace(-np.pi / 2, (3 * np.pi / 2), x1_num) #2 * half / x_step1)
             x2 = np.linspace(-np.pi / 2, (3 * np.pi / 2), x2_num) #2 * (1 - half) / x_step1)
